refactor: rename chatkit adapter modules
