+++ app/src/lib/runtime/codexChatKitAdapter.ts
