+++ app/src/lib/runtime/responsesDirectChatKitAdapter.ts
